Hm 2

+++ app/src/main/java/com/natan/shamilov/shmr25/domain/entity/Expense.kt
@@ -1,9 +1,9 @@
 package com.natan.shamilov.shmr25.domain.entity
 
 data class Expense(
-    val id: Int,
+    val id: Long,
     val category: Category,
-    val amount: Int,
+    val amount: Double,
     val comment: String = "",
     val createdAt: String
 ) {
